- Update documentation - Refactor code
