+++ labelit/mainapp/utils.py
@@ -4,7 +4,6 @@ import time
 
 from labelit.settings import BASE_DIR_PATH, LABELIT_DIRS
 
-
 class ErrorMessage():
     """Base class for labelit error messages"""
     def __init__(self, header="Error", message="Error occured", description=""):
@@ -18,14 +17,12 @@ def get_random_port():
         s.bind(('',0))
         return s.getsockname()[1]
 
-
 def get_label_studio_cmd(path_to_dataset, dataset_format, label_config_path, port):
     """Returns command to run label studio with given config"""
     command_template = "python {} start-multi-session --root-dir={} --input-path={} --input-format {} --sampling sequential -l {} -p {}"
     server_file = BASE_DIR_PATH / "flaskapp" / "server.py"
     command = command_template.format(server_file, LABELIT_DIRS['projects'], path_to_dataset, dataset_format, label_config_path, port)
     return command
-
 
 def start_tool_server(command):
     """Starts tool server, checks if it's running and return process ID"""
@@ -34,7 +31,6 @@ def start_tool_server(command):
     if p.poll() != None:
         raise Exception(f"Starting server failed with return code {p.returncode}")
     return p.pid
-
 
 def save_config_file(project_name, data):
     """Saves label studio config file"""
