+++ labelit/mainapp/views/__init__.py
@@ -1,3 +1,3 @@
-from mainapp.views.ls_proxy_view import LSProxyView 
+from mainapp.views.proxy_views import LSProxyView 
 from mainapp.views.project_views import create_project, list_projects, delete_view
 from mainapp.views.annotator_views import manage_annotators
